Feat/keg cli (#22)
* feat(core): add enabled to knut config schema

* test(cli): make the API testable

* feat(core): add keg init functionality

* feat(cli): add dir and init commands

* refactor(core): rename writeTo to toStorage

Rename `writeTo` to `toStorage` to keep things consistent.

* fix(core): write config.json when it is json

Writing config to the file system was incorrectly setting the file
extension to yaml instead of json when the contents is in json
formatting.

* feat: update date handling in knut-core

The code changes improve the way dates are handled in the knut-core package. The 'updated' field is now treated as a Date object instead of a string, and it's stringified only when necessary. This change provides more flexibility and accuracy when dealing with dates. Also, an optional 'now' parameter has been added to the 'updateContent' method to allow for custom timestamps.

* feat: Replace KegStorage with GenericStorage

The code changes replace the use of `KegStorage` with `GenericStorage`. This change simplifies the storage handling by using a more generic approach. The `Loader` type now returns a `Future<Optional<GenericStorage>>`, and all instances where `KegStorage.fromStorage(storage)` was used have been replaced with just returning the storage object.

Additionally, the method name in 'kegFile.ts' has been changed from 'writeTo' to 'toStorage'.

This commit also includes updates to function parameters across multiple files, replacing instances of `KegStorage` with `GenericStorage`.

* feat(core): add keg node writing and updating

Significant changes include:
- Simplified the backend by removing globalPlatform related functions
- Enhanced DexEntry to use Date type for 'updated' field instead of string
- Improved test coverage for keg.test.ts, including new tests for creating and editing nodes
- Added writeNodeContent function in keg.ts to handle content, updated time, and meta updates separately
- Updated KegNode class to handle date directly instead of converting it into a string
- Adjusted NodeStorage's utime method to work with Date objects directly.

* feat(cli): Enhance keg command and testing

Significant changes include:
- Refactored the KnutCommand function to streamline its return statement.
- Expanded test coverage for 'knut keg create' interface, including various input/output scenarios.
- Added a new 'create' action in the kegCli.ts file with corresponding CLI options and functionality.
- Modified the 'init' action to accommodate global options from parent commands.
- Fixed an issue in terminal.ts where data wasn't being pushed into the reader stream correctly.
- Updated knut.test.ts to clear localStorage after each test and added a new test case for creating a node on a specific keg.

* test(cli): made tests work in linux environment